day12 changes done.
