+++ lib/pages/cart_page.dart
@@ -69,17 +69,24 @@ class _CartListState extends State<_CartList> {
   final _cart = CartModel();
   @override
   Widget build(BuildContext context) {
-    return ListView.builder(
-        itemCount: _cart.items.length,
-        itemBuilder: (context, index) => ListTile(
-              leading: const Icon(Icons.done),
-              trailing: IconButton(
-                icon: const Icon(
-                  Icons.remove_circle_outline,
-                ),
-                onPressed: () {},
-              ),
-              title: _cart.items[index].name.text.make(),
-            ));
+    return _cart.items.isEmpty
+        ? "Nothing to Show!".text.xl3.makeCentered()
+        : ListView.builder(
+            itemCount: _cart.items.length,
+            itemBuilder: (context, index) => ListTile(
+                  leading: const Icon(Icons.done),
+                  trailing: IconButton(
+                    icon: const Icon(
+                      Icons.remove_circle_outline,
+                    ),
+                    onPressed: () {
+                      _cart.remove(_cart.items[index]);
+                      setState(() {
+                        
+                      });
+                    },
+                  ),
+                  title: _cart.items[index].name.text.make(),
+                ));
   }
 }
